Initial work on allowing multiple birdisle servers
Almost all calls to exit() have been replaced with exitFromServer(),
which calls pthread_exit() to terminate only the single thread.

The state is still global, so it doesn't actually work yet, but the
birdisle API has been redesigned around the concept. It's now also
possible to stop the server via the metasocket.