Make all the globals either const or __thread
This allows multiple servers to exist without clobbering each other.

Leave mutexes mostly alone, with the exception of moduleGIL which is
made __thread to avoid deadlocks as it locked as a rule rather than only
in critical sections (a thread-local mutex is an oxymoron, but it avoids
having to touch too much other code).

Modules that create threads will of course not work.