Free memory allocations manually
Remove the call to zmalloc_free_all(), and instead clean up individual
allocations. This commit frees all the allocations made by a trivial
startup-then-stop server, but may well leak resources that are only
allocated once data structures are manipulated.

The overheads in zmalloc to track the memory for zmalloc_free_all are
still in place, but will be removed later.

Also frees the allocated locale, which previously leaked.